More Scanner fixes and Parser fixes. - AST still needs rule names in the print out. - Scanner has line/col numbers that are incorrect… this was due to a bug in the STL size function. Now the scanner needs to be able to rewind correctly to get the right number. - Parser still need exceptions which allow it to continue parsing, despite errors… or it needs to default to panic mode to stop parsing. - Symbol table still needs to be added.

+++ src/Helper.hpp
@@ -12,13 +12,13 @@
 #include "Standard.hpp"
 
 static string format_error_lc(const string& type, const string& msg,
-                              const unsigned int line, const unsigned int column) {
+                              const unsigned long line, const unsigned long column) {
 	return string("[ " + type + ": " + msg + " @ " +
                   to_string(line) + ":" + to_string(column) + " ]" + '\n');
 }
 
 static void report_error_lc(const string& type, const string& msg,
-                            const unsigned int line, const unsigned int column) {
+                            const unsigned long line, const unsigned long column) {
     cerr << "[ " << type << ": " << msg << " @ " << line << ":" << column << " ]" << endl;
 }
